backend/requirements.txt
# Django and core dependencies
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# HTTP requests for AI services
requests==2.31.0

# Celery for background tasks
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.1.0

# Production server
gunicorn==21.2.0

# Development tools (optional)
django-debug-toolbar==4.2.0
